Update CKEditor version
The current packaged version of CKEditor (4.4.7) is over a year old now and has had 9 releases since then. Version 4.4.8 includes a security update and there are plenty of enhancements and bugfixes in the 4.5.x branch, which is now used in Drupal 8 core.

One particular bug is fixed in 4.5.6 regarding files embedded from the media_wysiwyg plugin. It's probably an edge case and not worth describing, unless I need to further justify upgrading CKEditor.